Always generate separate object files for liba{} and libso{}
While on some platforms they sometimes can be the same, they could also
be built differently (e.g., based on command line macros, etc). I guess
we could compare the set of options and if they are identical, then use
the same file. But that will complicate things quite a bit, so maybe in
version 2.