Allow back overriding variables specified in buildfiles
It is still not clear whether this is the right thing to allow, conceptually,
but with this disallowed it's hard to test this functionality. Perhaps we
should have an attribute [overridable]. The problem is one will also have to
set this variable to some value (e.g., [null]) which is not exactly the same
as undefined (especially when testing).