Redo handling of unhandled exceptions in async execution
Here is the problem: noexcept looses the call stack. That is, unlike an
unhandled exception, if noexcept is tripped, then you won't see the place
where it was thrown.

In this new implementation we now have noexcept only on the task thunk. And
the task is called via a thunk only in case of async execution. This means
that if we are executing serially (-j 1), then this will be an unhandled
exception, not noexcept. Hopefully will be a bit easier to debug.