Add support for for-loop
The semantics is similar to the C++11 range-based for:

list = 1 2 3
for i: $list
  print $i

Note that there is no scoping of any kind for the loop variable ('i' in
the above example).

See tests/loop/for.test for some examples/ideas.

In the future the plan is to also support more general while-loop as well
as break and continue.