Add {c,cxx}.class variables
Compiler class describes a set of compilers that follow more or less
the same command line interface. Compilers that don't belong to any of
the existing classes are in classes of their own (say, Sun CC would be
on its own if we were to support it).

Currently defined compiler classes:

gcc          gcc, clang, clang-apple, icc (on non-Windows)
msvc         msvc, clang-cl, icc (Windows)