Add support for relative to base scope command line variable overrides
Currently, if we say:

$ b dir/ ./foo=bar

The scope the foo=bar is set on is relative to CWD, not dir/. While this
may seem wrong at first, this is the least surprising behavior when we
take into account that there can be multiple dir/'s.

Sometimes, however, we do want the override directory to be treated relative
to (every) target's base scope that we are building. To support this we are
extending the '.' and '..' special directory names (which are still resolved
relative to CWD) with '...', which means "relative to the base scope of every
target in the buildspec". For example:

$ b dir/ .../foo=bar

Is equivalent to:

$ b dir/ dir/foo=bar

And:

$ b liba/ libb/ .../tests/foo=bar

Is equivalent to:

$ b liba/ libb/ liba/tests/foo=bar libb/tests/foo=bar