Add ability to disable automatic rpath, support for custom rpath-link
Specifically, the new config.bin.rpath.auto variable can be used to disable
automatic addition of prerequisite library rpaths, for example:

$ b config.bin.rpath.auto=false

Note that in this case rpath-link is still added where normally required
and for targets that support it (Linux and *BSD).

The new config.bin.rpath_link and config.bin.rpath_link.auto have the same
semantics as config.bin.rpath* but for rpath-link.