Add support for configuration exporting and importing
The new config.export variable specifies the alternative file to write the
configuration to as part of the configure meta-operation. For example:

$ b configure: proj/ config.export=proj-config.build

The config.export value "applies" only to the projects on whose root scope it
is specified or if it is a global override (the latter is a bit iffy but we
allow it, for example, to dump everything to stdout). This means that in order
to save a subproject's configuration we will have to use a scope-specific
override (since the default will apply to the outermost amalgamation). For
example:

$ b configure: subproj/ subproj/config.export=.../subproj-config.build

This could be somewhat unnatural but then it will be the amalgamation whose
configuration we normally want to export.

The new config.import variable specifies additional configuration files to be
loaded after the project's default config.build, if any. For example:

$ b create: cfg/,cc config.import=my-config.build

Similar to config.export, the config.import value "applies" only to the
project on whose root scope it is specified or if it is a global override.
This allows the use of the standard override "positioning" machinery (i.e.,
where the override applies) to decide where the extra configuration files are
loaded. The resulting semantics is quite natural and consistent with command
line variable overrides, for example:

$ b   config.import=.../config.build  # outermost amalgamation
$ b ./config.import=.../config.build  # this project
$ b  !config.import=.../config.build  # every project

Both config.export and config.import recognize the special `-` file name as an
instruction to write/read to/from stdout/stdin, respectively. For example:

$ b configure: src-prj/ config.export=- | b configure: dst-prj/ config.import=-