Adjust variable block applicability in dependency chains
Before the block used to apply to the set of prerequisites before the last
`:`. This turned out to be counterintuitive and not very useful since
prerequisite-specific variables are a lot less common than target specific.
And it doesn't fit with ad hoc recipes.

The new rule is if the chain ends with `:`, then the block applies to the last
set of prerequisites. Otherwise, it applies to the last set of targets. For
example:

./: exe{test}: cxx{main}
{
  test = true        # Applies to the exe{test} target.
}

./: exe{test}: libue{test}:
{
  bin.whole = false  # Applies to the libue{test} prerequisite.
}

This is actually consistent with both non-chain and non-block cases.
Consider:

exe{test}: cxx{main}
{
  test = true
}

exe{test}: libue{test}:
{
  bin.whole = false
}

exe{test}: libue{test}: bin.whole = false

The only exception we now have in this overall approach of "if the
dependency declaration ends with a colon, then what follows is for a
prerequisite" is for the first semicolon:

exe{test}:
{
  test = true
}

exe{test}: test = true

But that's probably intuitive enough since there cannot be a prerequisite
without a target.