Add support for project-local importation
An import without a project name or with the same name as the importing
project's is now treated as importation from the same project.

For example, given the libhello project that exports the lib{hello} target, a
buildfile for an executable in the same project instead of doing something
like this:

include ../libhello/
exe{hello}: ../libhello/lib{hello}

Can now do this:

import lib = libhello%lib{hello}

Or:

import lib = lib{hello}

And then:

exe{hello}: $lib

Note that a target in project-local importation must still be exported in
the project's export stub. In other words, project-local importation goes
through the same mechanisms as normal import.