Initial attempt (gzip -n and preserve post-processed manifest timestamp)
Turns our this is not enough since tar saves timestamps (as well as owner
information, file order, etc) including for directories and getting rid of
that in a portable way is not going to be easy. Looks like the most plausible
way would be to switch to embedded libarchive (provided it's possible to
achieve everything we need).