Make libul{} rule match without hint
In the end, the extra jumping through the hoops doesn't justify the extra
safety we gain. The only plausible accidental mistake is making libul{} a
dependency of ./ but then we don't prevent the same for libue{}, which also
doesn't make much sense. Though, the consequences of doing this for libul{}
could be more severe, like messed up for-install'ness. Oh, well, I guess
people will just have to pay attention (this could be a good check for the
linter we've been thinking about).