Switch to using libpkg-config instead of libpkgconf by default
The use of (now deprecated) libpkgconf is still possible by setting
config.build2.libpkgconf to true. Note that libpkgconf is known to
have issues on Windows and Mac OS so this should only be used on
Linux and maybe BSDs. Also note that we will only keep this until
upstream (again) breaks backwards compatibility at which point we
will drop this support.