Allow empty prerequisites in prerequisite-specific variable assignment/block
The old semantics was not very consistent, consider:

exe{foo}: cxx{$empty}                  # Ok.
exe{foo}: cxx{$empty}: include = false # Not ok?

So now both are ok, as well as the block variant:

exe{foo}: cxx{$empty}:
{
  include = false
}

Note that the empty prerequisite list in the dependency chain is still an
error:

./: exe{$empty}: cxx{foo} # Error.

Note also that the syntactically-empty prerequisite list was and still is
an error:

exe{foo}: : include = false # Error.