Add ability to remap paths in distribution
Specifically, the dist target-specific variable now can specify a path
besides true or false. This path is the "imaginary" source location which
is used to derive the corresponding distribution local. This location can
be either a directory path (to remap with the same file name) or a file
path (to remap with a different name). If the path is relative, then it's
treated relative to the target directory. Note that to make things less
error prone, simple paths without any directory separators are not allowed
(use ./<name> instead).

Note that if multiple targets end up with the same source location, the
behavior is undefined and no diagnostics is issued.

Note also that such remapping has no effect in the bootstrap distribution
mode.