Remap Apple Clang 14.0.3 and 15.0.0 to vanilla Clang 14.0 and 15.0
Lately, we started seeing __LIBCPP_VERSION values like 15.0.6 or 16.0.2 which
would suggest the base is 15.0.5 or 16.0.1. But that assumption did not check
out with the actual usage. For example, vanilla Clang 16 should no longer
require -fmodules-ts but the Apple's version (that is presumably based on it)
still does. So the theory here is that Apple upgrades to newer libc++ while
keeping the old compiler. Which means we must be more conservative and assume
something like 15.0.6 is still 14-based.