Fall back to amalgamation manifest in version module
Note that in this case, currently, we ignore everything except verifying
the build2 version constraint and deriving the default script syntax
version based on it.

The idea is to be able to load the version module in subprojects (such
as tests/) and then have their standalone build derive the script syntax
version consistent with when built as part of the amalgamation.