Add config.build2.default_{c,cxx}, config.build2.host_{c,cxx} variables
The first set allows specifying the default C and C++ compilers. These are
used as the default values for config.c and config.cxx, respectively, if
not specified explicitly by the user.

The second set allows specifying the host C and C++ compilers. These are
used as the values for config.c and config.cxx, respectively, in embedded
~host and ~build2 configurations.