Detect and diagnose cases where we should not switch to another project
In particular, we now don't allow switching to a different project (or
no project) when assigning scope/target/pattern-specific variables. We
also disallow it when explicitly switching the scope except in export
stubs. One can still switch to any project (but not to no project) with
include.

Note: fixes GH issues #475 and #469.